Rat in a maze-Backtracking
In this, basically, when we visit a cell, we set its value to zero so that while running its dfs, we don't revisit the already visited node. This also helps in cycle detection because we don't revisit the same node while backtracking. The order is D,L,R,U and thus the function calls are also in that order.
We can also have a visited array initially initialised to all false for all cells and then only visiting the nodes that are false. Once a vertex is visited, it is not visited again unless it is set to true during backtracking. Thus, cycles are not encountered.